yatube/posts/models.py
from django.contrib.auth import get_user_model
from django.db import models
from pytils.translit import slugify

User = get_user_model()


class Group(models.Model):
    title = models.CharField(
        max_length=200,
        verbose_name='group name',
        help_text='what should we call it')
    slug = models.SlugField(
        max_length=200,
        unique=True,
        verbose_name='group url',
        help_text='give it a nice address')
    description = models.TextField(
        verbose_name='Описание',
        max_length=200,
        help_text='what is it about?')

    def __str__(self):
        return self.title

    def save(self, *args, **kwargs):
        if not self.slug:
            self.slug = slugify(self.title)[:100]
        super().save(*args, **kwargs)


class Post(models.Model):
    text = models.TextField(
        verbose_name='Текст поста:',
        help_text='Пишите здесь всё, что хотите')
    pub_date = models.DateTimeField(
        verbose_name='Дата публикации',
        auto_now_add=True)
    author = models.ForeignKey(
        User,
        on_delete=models.CASCADE,
        related_name='posts'
    )
    group = models.ForeignKey(
        Group,
        on_delete=models.SET_NULL,
        related_name='posts',
        verbose_name='Выберите группу:',
        help_text='Опубликовать этот пост в группе?',
        blank=True,
        null=True
    )

    class Meta:
        ordering = ['-pub_date']

    def __str__(self):
        return self.text[:15]
